+++ Search/LocalizationStrategyInterface.php
@@ -20,18 +20,31 @@ interface LocalizationStrategyInterface
      * Provide localized version of the given index name.
      *
      * @param string $indexName
+     * @param string $locale
+     *
+     * @return
      */
     public function localizeIndexName($indexName, $locale);
+
+    /**
+     * Remove the localization part of the given index name.
+     *
+     * @param $indexName
+     *
+     * @return mixed
+     */
+    public function delocalizeIndexName($indexName);
 
     /**
      * Check to see if a given variant name is a variant of the
      * given indexName. This is used to determine which indexes can
      * be grouped together (e.g. when purging an index).
      *
-     * @param string $baseIndexName
-     * @param string $candidateIndexName
+     * @param string $indexName
+     * @param string $variantName
+     * @param string $locale
      *
      * @return bool
      */
-    public function isIndexVariantOf($indexName, $variantName);
+    public function isIndexVariantOf($indexName, $variantName, $locale = null);
 }