+++ Tests/Unit/Search/Metadata/Driver/XmlDriverTest.php
@@ -13,6 +13,7 @@ namespace Massive\Bundle\SearchBundle\Search\Metadata\Driver;
 
 use Massive\Bundle\SearchBundle\Search\Factory;
 use Massive\Bundle\SearchBundle\Search\Metadata\ClassMetadata;
+use Massive\Bundle\SearchBundle\Search\Metadata\Field\Expression;
 use Massive\Bundle\SearchBundle\Search\Metadata\Field\Field;
 use Massive\Bundle\SearchBundle\Search\Metadata\Field\Property;
 use Massive\Bundle\SearchBundle\Search\Metadata\IndexMetadata;
@@ -45,26 +46,32 @@ class XmlDriverTest extends \PHPUnit_Framework_TestCase
     {
         $this->factory = $this->prophesize(Factory::class);
         $this->fileLocator = $this->prophesize(FileLocatorInterface::class);
+
         $this->reflectionClass = $this->prophesize(\ReflectionClass::class);
+        $this->reflectionClass->getName()->willReturn('Sulu\Bundle\ExampleBundle\Entity\Example');
 
         $that = $this;
 
-        $this->factory->createClassMetadata(Argument::cetera())->will(function($arguments) use($that) {
+        $this->factory->createClassMetadata(Argument::cetera())->will(function ($arguments) use ($that) {
             $classMetadata = new ClassMetadata($that->reflectionClass->getName());
 
             return $classMetadata;
         });
 
-        $this->factory->createMetadataProperty(Argument::cetera())->will(function($arguments) use($that) {
+        $this->factory->createMetadataProperty(Argument::cetera())->will(function ($arguments) use ($that) {
             return new Property($arguments[0]);
         });
 
-        $this->factory->createMetadataField(Argument::cetera())->will(function($arguments) use($that) {
+        $this->factory->createMetadataField(Argument::cetera())->will(function ($arguments) use ($that) {
             return new Field($arguments[0]);
         });
 
-        $this->factory->createIndexMetadata()->will(function() use($that) {
+        $this->factory->createIndexMetadata()->will(function () use ($that) {
             return new IndexMetadata();
+        });
+
+        $this->factory->createMetadataExpression(Argument::cetera())->will(function ($arguments) use ($that) {
+            return new Expression($arguments[0]);
         });
 
         $this->xmlDriver = new XmlDriver($this->factory->reveal(), $this->fileLocator->reveal());
@@ -72,8 +79,6 @@ class XmlDriverTest extends \PHPUnit_Framework_TestCase
 
     public function testLoadMetadataFromFile()
     {
-        $this->reflectionClass->getName()->willReturn('Sulu\Bundle\ExampleBundle\Entity\Example');
-
         $metadata = $this->xmlDriver->loadMetadataFromFile(
             $this->reflectionClass->reveal(),
             __DIR__ . '/../../../../Resources/DataFixtures/Mapping/Example.xml'
@@ -84,7 +89,7 @@ class XmlDriverTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('id', $indexMetadata->getIdField()->getProperty());
         $this->assertEquals('title', $indexMetadata->getTitleField()->getProperty());
         $this->assertEquals('description', $indexMetadata->getDescriptionField()->getProperty());
-        $this->assertEquals('example', $indexMetadata->getIndexName());
+        $this->assertEquals('example', $indexMetadata->getIndexName()->getName());
         $this->assertEquals('example', $indexMetadata->getCategoryName());
         $this->assertEquals('locale', $indexMetadata->getLocaleField()->getProperty());
 
@@ -93,5 +98,17 @@ class XmlDriverTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('title', $fieldMappings['title']['field']->getName());
         $this->assertEquals('string', $fieldMappings['description']['type']);
         $this->assertEquals('description', $fieldMappings['description']['field']->getName());
+    }
+
+    public function testLoadMetadataFromFileWithEvaluatingIndex()
+    {
+        $metadata = $this->xmlDriver->loadMetadataFromFile(
+            $this->reflectionClass->reveal(),
+            __DIR__ . '/../../../../Resources/DataFixtures/Mapping/ExampleWithEvaluatingIndex.xml'
+        );
+
+        $indexMetadata = $metadata->getIndexMetadata('_default');
+
+        $this->assertEquals('object.getWebspace()', $indexMetadata->getIndexName()->getExpression());
     }
 }