+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+dev-develop
+-----------
+
+### Feature
+
+- [Configuration] Index name can now also be created using expression language
+
 0.8.0
 -----------
 
@@ -16,7 +23,7 @@ CHANGELOG
 - [Metadata] Added scalar check
 
 0.6.0
------------
+-----
 
 ### Bugfix